+++ sandbox.controller/src/kz/greetgo/sandbox/controller/model/ClientRecords.java
@@ -2,10 +2,6 @@ package kz.greetgo.sandbox.controller.model;
 
 import java.util.ArrayList;
 import java.util.List;
-import kz.greetgo.sandbox.controller.model.Address;
-import kz.greetgo.sandbox.controller.model.Charm;
-import kz.greetgo.sandbox.controller.model.Gender;
-import kz.greetgo.sandbox.controller.model.PhoneNumber;
 
 public class ClientRecords {
   public String id;
@@ -18,7 +14,7 @@ public class ClientRecords {
   public Address addressF;
   public Address addressR;
   public List<PhoneNumber> phoneNumbers = new ArrayList<>();
-  public int totalBalance;/**/
-  public int minBalance;/**/
-  public int maxBalance;/**/
+  public double totalBalance;/**/
+  public double minBalance;/**/
+  public double maxBalance;/**/
 }
